Use a WebWorker for timers rather than the `window` ones
This is mostly helpful for development: setTimeout / setInterval events
on tabs which are not in focus are constrained to fire at most once per
second, which makes the deal unbearably slow. However, WebWorkers are
not limited in this way, so we can use the worker's timer to implement
the frontend's timer.