Fix race in healthcheck callback
Fix race where the old callback and the new callback might interleave,
causing the healthcheck timer to unexpectedly fire. Preventing the timer
from firing is complex, especially since the setTimeout call is actually
running in a web worker, but we know that the healthcheck is only valid
if it's still the most recent timer, so we can resolve the user-facing
issue.