Add a "meta-level" concept and wrap ranks after A
Some people like to keep playing after the A rank. In theory, it would
probably be best to have an "NT" rank, but as an intermediary step,
let's allow players to keep track of an ephemeral "meta-level". Note:
there is no mechanism currently to _set_ the meta-level, though it
wouldn't be a big deal to add one.