Use a canvas hack to constrain the bounding box of a card
If you use a canvas element, apparently you can get the cards to be
correctly bounded with comparatively limited hacks on the actual CSS
side. This is kind of terrible, especially the bit where we make a
temporary span just to grab a value out of the CSS file. But... it seems
to work?

tmp

try to fix on Windows

Handle device pixel ratio for hidpi computers

try to line things up; set dpr to 1.5x the real dpr

update offset a bit