Tighten card boundaries using Canvas tricks (#192)
* Use a canvas hack to constrain the bounding box of a card

If you use a canvas element, apparently you can get the cards to be
correctly bounded with comparatively limited hacks on the actual CSS
side. This is kind of terrible, especially the bit where we make a
temporary span just to grab a value out of the CSS file. But... it seems
to work?

* Use SVG for rendering rather than canvas to help deal with pixel-ratio issues