Use WASM to sort and group cards from the hands structure (#207)
Rather than encoding cards twice (in a fairly expensive way!) on the
wire, use code in the frontend to generate the appropriate list-of-cards
format. Previously, we didn't do this because the sort order of the
cards was undefined without additional state; now, we can do better by
using logic in WASM to sort.